Updateinventory

+++ src/main/java/com/jsp/wms/repository/InventoryRepository.java
@@ -5,5 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jsp.wms.entity.Inventory;
 
-public interface InventoryRepository extends JpaRepository<Inventory, Long> {
+
+public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
+
 }
